+++ src/pages/my-loans/Lending.js
@@ -1,0 +1,40 @@
+import { useState, useEffect } from 'react';
+
+// material-ui
+import { Grid } from '@mui/material';
+
+// ant design
+import { Table } from 'antd';
+
+// project import
+import { useAuth } from 'pages/authentication/auth-forms/AuthProvider';
+
+// react
+import { useNavigate } from 'react-router-dom';
+
+const Lending = () => {
+    const [dataLoading, setLoading] = useState(false);
+    const [items, setItems] = useState([]);
+    const { user, loading } = useAuth();
+
+    useEffect(() => {
+        load_endpoint(
+            'http://127.0.0.1:8000/loans/user/self/accepted?perspective=lender&recent=True',
+            (result) => {
+                // setItems(result);
+                setLoading(false);
+            },
+            (error) => {
+                setLoading(false);
+            }
+        );
+    }, []);
+
+    return (
+        <div>
+            <Table columns={columns} dataSource={[]} /> {/* items */}
+        </div>
+    );
+};
+
+export default Lending;
